Move new classes to product_types.py
